Refactor nodes into separate files
