+++ app/javascript/controllers/tiptap_controller.js
@@ -1,160 +1,15 @@
 import { Controller } from '@hotwired/stimulus'
 import { debounce } from 'lodash/function'
-import { Editor, Node, mergeAttributes } from '@tiptap/core'
+import { Editor } from '@tiptap/core'
 import StarterKit from '@tiptap/starter-kit'
 import Underline from '@tiptap/extension-underline'
 import Placeholder from '@tiptap/extension-placeholder'
 import Link from '@tiptap/extension-link'
 import Image from '@tiptap/extension-image'
 
-const DonationGoalNode = Node.create({
-  name: 'donationGoal',
-  group: 'block',
-  priority: 2000,
-  renderHTML({ HTMLAttributes }) {
-    return [
-      'div',
-      mergeAttributes(HTMLAttributes, {
-        class:
-          'donationGoal relative card shadow-none border flex flex-col py-2 my-2',
-      }),
-      [
-        'p',
-        { class: 'text-center italic' },
-        'Your progress towards your goal will display here',
-      ],
-      [
-        'div',
-        { class: 'bg-gray-200 dark:bg-neutral-700 rounded-full w-full' },
-        [
-          'div',
-          {
-            class:
-              'h-full bg-primary rounded w-1/2 flex items-center justify-center',
-          },
-          ['p', { class: 'text-sm text-black p-[1px] my-0' }, '50%'],
-        ],
-      ],
-    ]
-  },
-  parseHTML() {
-    return [
-      {
-        tag: 'div',
-        getAttrs: node => node.classList.contains('donationGoal') && null,
-      },
-    ]
-  },
-  addCommands() {
-    return {
-      addDonationGoal:
-        () =>
-        ({ commands }) => {
-          return commands.insertContent({ type: this.name })
-        },
-    }
-  },
-})
-
-const HcbCodeNode = Node.create({
-  name: 'hcbCode',
-  group: 'block',
-  priority: 2000,
-  addAttributes() {
-    return {
-      code: {},
-    }
-  },
-  renderHTML({ HTMLAttributes }) {
-    return [
-      'div',
-      mergeAttributes(HTMLAttributes, {
-        class:
-          'hcbCode relative card shadow-none border flex flex-col py-2 my-2',
-      }),
-      [
-        'p',
-        { class: 'italic text-center' },
-        `Your transaction (${HTMLAttributes.code}) will appear here.`,
-      ],
-    ]
-  },
-  parseHTML() {
-    return [
-      {
-        tag: 'div',
-        getAttrs: node => node.classList.contains('hcbCode') && null,
-      },
-    ]
-  },
-  addCommands() {
-    return {
-      addHcbCode:
-        code =>
-        ({ commands }) => {
-          return commands.insertContent({
-            type: this.name,
-            attrs: { code },
-          })
-        },
-    }
-  },
-})
-
-const DonationSummaryNode = Node.create({
-  name: 'donationSummary',
-  group: 'block',
-  priority: 2000,
-  addAttributes() {
-    return {
-      startDate: {},
-    }
-  },
-  renderHTML({ HTMLAttributes }) {
-    let start
-    if (HTMLAttributes.startDate) {
-      start = new Date(HTMLAttributes.startDate)
-    } else {
-      const date = new Date()
-      const currentMonth = date.getMonth()
-      date.setMonth(currentMonth - 1)
-      if (date.getMonth() == currentMonth) date.setDate(0)
-      date.setHours(0, 0, 0, 0)
-
-      start = date
-    }
-
-    return [
-      'div',
-      mergeAttributes(HTMLAttributes, {
-        class:
-          'donationSummary relative card shadow-none border flex flex-col py-2 my-2',
-      }),
-      [
-        'p',
-        { class: 'italic text-center' },
-        `A donation summary starting on ${start.toDateString()} will appear here.`,
-      ],
-    ]
-  },
-  parseHTML() {
-    return [
-      {
-        tag: 'div',
-        getAttrs: node => node.classList.contains('donationSummary') && null,
-      },
-    ]
-  },
-  addCommands() {
-    return {
-      addDonationSummary:
-        () =>
-        ({ commands }) => {
-          return commands.insertContent({ type: this.name })
-        },
-    }
-  },
-})
+import { DonationGoalNode } from './tiptap/nodes/donation_goal_node'
+import { HcbCodeNode } from './tiptap/nodes/hcb_code_node'
+import { DonationSummaryNode } from './tiptap/nodes/donation_summary_node'
 
 export default class extends Controller {
   static targets = ['editor', 'form', 'contentInput', 'autosaveInput']
